+++ app/app.py
@@ -4,10 +4,10 @@
 from sys import path
 from typing import Literal
 
-from .helper.get_and_configure_logger import debug_on_global, get_and_configure_logger
-from .helper.get_and_configure_system_info import debug_system_info, show_sysinfo_global
-from .helper.parse_configs_into_paramdict import get_param_dict
-from .helper.prepare_ml_input import PipelineOutput, prepare_pipeline
+from .utils.get_and_configure_logger import debug_on_global, get_and_configure_logger
+from .utils.get_and_configure_system_info import debug_system_info, show_sysinfo_global
+from .utils.parse_configs_into_paramdict import get_param_dict
+from .utils.prepare_ml_input import PipelineOutput, prepare_pipeline
 
 if debug_on_global or show_sysinfo_global:
     logger = get_and_configure_logger(__name__)